fix(ci): python-coverage-shards (8) — split pre_push_validation to clear file-size guardrail
The maintainability guardrail test
(test_first_party_code_files_stay_under_line_limit) failed because
src/awf/runtime/pr_monitor_runner/pre_push_validation.py reached 1502
lines, 2 over the 1,500-line first-party limit, after the pre-push pause
work landed.

Extract the pure failure-classification helpers (the ~150-line group
operating on ValidationResult with no `self` dependency) into a cohesive
sibling module pre_push_validation_failures.py and import the names the
orchestration module still uses. This keeps the helpers single-sourced
(just in a more focused module) and brings pre_push_validation.py to
1364 lines. Update the structural single-source test to assert the
helpers now live exactly once in the failures module and not in the
orchestrator, and repoint test references accordingly.
