+++ src/awf/runtime/pr_monitor_runner/pre_push_validation.py
@@ -20,7 +20,6 @@ from awf.control.executor.helpers import (
     _should_run_local_coverage,
     _validation_run_command_records,
     _validation_run_coverage_metadata,
-    _validation_run_reason_code,
     _validation_tier_for_workspace,
 )
 from awf.control.executor.logging_ops import _validation_run_log_stream_refs
@@ -43,6 +42,13 @@ from awf.runtime.pr_monitor_runner.pre_push_validation_constants import (
     _PRE_PUSH_VALIDATION_REPARENT_FAILED_REASON,
     _PRE_PUSH_VALIDATION_ROLLBACK_FAILED_REASON,
     _PRE_PUSH_VALIDATION_TOOLCHAIN_MISSING_REASON,
+)
+from awf.runtime.pr_monitor_runner.pre_push_validation_failures import (
+    _failed_pre_push_commands,
+    _pre_push_validation_reason_code_for_preferred_failure,
+    _preferred_pre_push_failure,
+    _preferred_pre_push_failure_from_failures,
+    _pure_toolchain_missing_failure_for_result,
 )
 from awf.runtime.pr_monitor_runner.remote_ops import _GitPushResult
 from awf.runtime.validation import profile_phase_command_plan
@@ -79,157 +85,6 @@ PRE_PUSH_VALIDATION_TOOLCHAIN_MISSING_REASON = _PRE_PUSH_VALIDATION_TOOLCHAIN_MI
 PRE_PUSH_VALIDATION_REPARENT_FAILED_REASON = _PRE_PUSH_VALIDATION_REPARENT_FAILED_REASON
 
 _FAILING_COMMAND_DETAIL_LIMIT = 1000
-
-
-def _failed_pre_push_commands(result: ValidationResult) -> tuple[ValidationCommandResult, ...]:
-    """Return failed command-like records from a validation result."""
-    failures: list[ValidationCommandResult] = []
-    if result.migration is not None and not result.migration.ok:
-        failures.append(result.migration)
-    failures.extend(command for command in result.commands if command.blocks_validation)
-    coverage_command = result.coverage.command_result if result.coverage is not None else None
-    if coverage_command is not None and not coverage_command.ok:
-        failures.append(coverage_command)
-    return tuple(failures)
-
-
-def _first_real_pre_push_failure(result: ValidationResult) -> ValidationCommandResult | None:
-    """Return the first non-127 failure, giving real lint/test failures precedence."""
-    failures = _failed_pre_push_commands(result)
-    return _first_real_pre_push_failure_for_result(result, failures)
-
-
-def _first_failure_outside_collected_failures(
-    result: ValidationResult,
-    failures: tuple[ValidationCommandResult, ...],
-) -> ValidationCommandResult | None:
-    """Return ``first_failure`` when it is not represented in collected commands."""
-    first_failure = result.first_failure
-    if first_failure is None:
-        return None
-    if any(first_failure is failure for failure in failures):
-        return None
-    return first_failure
-
-
-def _first_real_pre_push_failure_for_result(
-    result: ValidationResult,
-    failures: tuple[ValidationCommandResult, ...],
-) -> ValidationCommandResult | None:
-    """Return the first non-127 failure across command and provider records.
-
-    Provider-level ``first_failure`` may describe a policy failure whose
-    underlying command succeeded, such as coverage below threshold with
-    ``ok=True`` and ``returncode=0``.
-    """
-    real_failure = _first_real_pre_push_failure_from_failures(failures)
-    if real_failure is not None:
-        return real_failure
-    first_failure = _first_failure_outside_collected_failures(result, failures)
-    if first_failure is not None and first_failure.returncode != 127:
-        return first_failure
-    return None
-
-
-def _first_real_pre_push_failure_from_failures(
-    failures: tuple[ValidationCommandResult, ...],
-) -> ValidationCommandResult | None:
-    """Return the first non-127 failure from an already collected failure tuple."""
-    return next(
-        (failure for failure in failures if failure.returncode != 127),
-        None,
-    )
-
-
-def _pure_toolchain_missing_failure(
-    result: ValidationResult,
-) -> ValidationCommandResult | None:
-    """Return the first 127 failure only when all command failures are command-not-found.
-
-    Mixed results are treated as genuine validation failures so a real lint/test
-    failure is not hidden behind an earlier missing-tool diagnostic.
-    """
-    failures = _failed_pre_push_commands(result)
-    return _pure_toolchain_missing_failure_for_result(result, failures)
-
-
-def _pure_toolchain_missing_failure_for_result(
-    result: ValidationResult,
-    failures: tuple[ValidationCommandResult, ...],
-) -> ValidationCommandResult | None:
-    """Return a pure 127 failure, including command-less provider failures."""
-    first_failure = _first_failure_outside_collected_failures(result, failures)
-    if first_failure is not None and first_failure.returncode != 127:
-        return None
-    toolchain_failure = _pure_toolchain_missing_failure_from_failures(failures)
-    if toolchain_failure is not None:
-        return toolchain_failure
-    if failures:
-        return None
-    if first_failure is not None and first_failure.returncode == 127:
-        return first_failure
-    return None
-
-
-def _pure_toolchain_missing_failure_from_failures(
-    failures: tuple[ValidationCommandResult, ...],
-) -> ValidationCommandResult | None:
-    """Return the first 127 failure only when the collected failures are all 127."""
-    if not failures:
-        return None
-    if any(failure.returncode != 127 for failure in failures):
-        return None
-    return failures[0]
-
-
-def _preferred_pre_push_failure(result: ValidationResult) -> ValidationCommandResult | None:
-    """Return the failure that should drive diagnostics and repair prompts."""
-    return _preferred_pre_push_failure_from_failures(
-        result,
-        _failed_pre_push_commands(result),
-    )
-
-
-def _preferred_pre_push_failure_from_failures(
-    result: ValidationResult,
-    failures: tuple[ValidationCommandResult, ...],
-) -> ValidationCommandResult | None:
-    """Return the preferred failure using an already collected failure tuple."""
-    real_failure = _first_real_pre_push_failure_for_result(result, failures)
-    if real_failure is not None:
-        return real_failure
-    toolchain_failure = _pure_toolchain_missing_failure_from_failures(failures)
-    if toolchain_failure is not None:
-        return toolchain_failure
-    return result.first_failure
-
-
-def _pre_push_validation_reason_code_for_preferred_failure(
-    result: ValidationResult,
-    preferred_failure: ValidationCommandResult | None,
-) -> str:
-    """Return the validation reason for an already selected preferred failure."""
-    validation_reason_code = _validation_run_reason_code(result)
-    if preferred_failure is None:
-        return validation_reason_code
-    coverage_command = result.coverage.command_result if result.coverage is not None else None
-    # ValidationResult.first_failure returns this same coverage command object when
-    # a coverage policy fails; preserve that identity if coverage results are copied.
-    if (
-        result.coverage is not None
-        and not result.coverage.ok
-        and preferred_failure is coverage_command
-    ):
-        return validation_reason_code
-    return preferred_failure.reason_code
-
-
-def _pre_push_validation_reason_code(result: ValidationResult) -> str:
-    """Return the underlying validation reason, honoring mixed-failure precedence."""
-    return _pre_push_validation_reason_code_for_preferred_failure(
-        result,
-        _preferred_pre_push_failure(result),
-    )
 
 
 def _safe_pre_push_validation_artifact_name(value: str) -> str:
